# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Plant API",
        "description": "Ugne's Simple Plant API",
        "contact": {
            "email": "n00201661@student.iadt.ie"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/plants": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Display a listing of the resource.",
                "description": "Displays all plants",
                "operationId": "6e2a3a91967b1e364c87b755cd5152f6",
                "responses": {
                    "200": {
                        "description": "Successful operation, Returns a list of Plants in JSON format"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Plants"
                ],
                "summary": "Create a new Plant",
                "description": "Stores the plant in the DB",
                "operationId": "storePlant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category",
                                    "origin",
                                    "climate_id",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample Name"
                                    },
                                    "categories": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "[1, 2]"
                                    },
                                    "climate_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "3"
                                    },
                                    "origin": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "East Asia"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A description about this lovely plant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/plants/{id}": {
            "get": {
                "tags": [
                    "Plants"
                ],
                "summary": "Display the specified resource.",
                "description": "Gets a plant by ID",
                "operationId": "5f7b2efcf3ae43a0485b376c550bcdbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Plants"
                ],
                "summary": "Update a plant",
                "description": "Updates plants in the database",
                "operationId": "putPlant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Plant id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "categories",
                                    "climate_id",
                                    "origin",
                                    "description"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Sample name"
                                    },
                                    "categories": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "[3, 5]"
                                    },
                                    "climate_id": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "8"
                                    },
                                    "origin": {
                                        "type": "string",
                                        "format": "integer",
                                        "example": "East Asia"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "A long description about this lovely plant"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": ""
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Plants"
                ],
                "summary": "Delete a Plant",
                "description": "Delete Plant",
                "operationId": "destroyPlant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of a Plant",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": "204"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}